requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
tiktoken
python-dotenv
langchain-community
langchain-google-genai
gradio